Implement Stack
Lesson learned.

If you want to define a method on a struct that mutates the struct, then you must define the method on the _pointer_ to the struct.

e.g.:

// Bad 🙅‍♀️
func (l LinkedList) Append(val int) {}

// Good 💯
func (l *LinkedList) Append(val int) {}

I'm not entirely sure why this is yet.

When you define the method on the struct literal, does calling the method pass just the _value_ (i.e.,, a copy of the struct)?

That could make sense. Then if you wanted to write more functional code, you could always define your methods on the struct literal, and return a new struct.

When you define the method on a pointer to the struct, then you are modifying the particular instance that the caller is pointing to.